move work with vacancy to repository
