+++ app/src/main/java/ru/practicum/android/diploma/network/data/NetworkClient.kt
@@ -1,18 +1,61 @@
 package ru.practicum.android.diploma.network.data
 
-import ru.practicum.android.diploma.network.data.dto.requests.VacanciesFilter
-import ru.practicum.android.diploma.network.data.dto.response.FilterArea
-import ru.practicum.android.diploma.network.data.dto.response.FilterIndustry
-import ru.practicum.android.diploma.network.data.dto.response.VacancyDetail
-import ru.practicum.android.diploma.network.data.dto.response.VacancyResponse
+import android.util.Log
+import com.google.gson.JsonSyntaxException
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.HttpException
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import ru.practicum.android.diploma.App.Companion.getAppContext
+import ru.practicum.android.diploma.BuildConfig
+import ru.practicum.android.diploma.utils.NetworkUtils
 
-interface NetworkClient {
+class NetworkClient {
 
-    suspend fun getAreas(): ApiResult<List<FilterArea>>
+    companion object {
 
-    suspend fun getIndustries(): ApiResult<List<FilterIndustry>>
+        private const val BASE_URL = "https://practicum-diploma-8bc38133faba.herokuapp.com/"
 
-    suspend fun getVacancy(id: String): ApiResult<VacancyDetail>
+        private const val HTTP_CODE_400 = 400
+    }
 
-    suspend fun getVacancies(vacanciesFilter: VacanciesFilter): ApiResult<VacancyResponse>
+    private val networkUtils = NetworkUtils(getAppContext())
+
+    private val client = OkHttpClient.Builder()
+        .retryOnConnectionFailure(true)
+        .addInterceptor(HttpLoggingInterceptor().apply { level = HttpLoggingInterceptor.Level.BODY })
+        .addInterceptor { chain ->
+            val newRequest = chain.request().newBuilder()
+                .addHeader("Authorization", "Bearer ${BuildConfig.API_ACCESS_TOKEN}")
+                .addHeader("Accept", "application/json")
+                .build()
+            chain.proceed(newRequest)
+        }
+        .build()
+
+    val retrofit = Retrofit.Builder()
+        .baseUrl(BASE_URL)
+        .client(client)
+        .addConverterFactory(GsonConverterFactory.create())
+        .build()
+
+    suspend fun <T> doRequest(serviceRequest: suspend () -> T): ApiResult<T> {
+        if (networkUtils.isInternetAvailable() == false) {
+            return ApiResult.NoInternetConnection
+        }
+        return withContext(Dispatchers.IO) {
+            try {
+                val response = serviceRequest()
+                ApiResult.Success(response)
+            } catch (e: HttpException) {
+                ApiResult.Error(e.code())
+            } catch (e: JsonSyntaxException) {
+                Log.e("NetworkClient", "JSON parsing error", e)
+                ApiResult.Error(HTTP_CODE_400)
+            }
+        }
+    }
 }
